+++ src/zephyr/client.ts
@@ -19,6 +19,7 @@ import { CreateTestCycle } from "./tool/test-cycle/create-test-cycle";
 import { GetTestCycle } from "./tool/test-cycle/get-test-cycle";
 import { GetTestCycles } from "./tool/test-cycle/get-test-cycles";
 import { UpdateTestCycle } from "./tool/test-cycle/update-test-cycle.ts";
+import { CreateTestExecution } from "./tool/test-execution/create-test-execution";
 import { GetTestExecution } from "./tool/test-execution/get-test-execution";
 import { GetTestExecutions } from "./tool/test-execution/get-test-executions";
 
@@ -82,6 +83,7 @@ export class ZephyrClient implements Client {
       new CreateTestCycle(this),
       new UpdateTestCase(this),
       new UpdateTestCycle(this),
+      new CreateTestExecution(this),
     ];
 
     for (const tool of tools) {